@@ -1,11 +1,11 @@
 name: sqlite3_library_linux
 description: Flutter package to bundle sqlite3 library to your linux apps
-version: 1.0.0
+version: 2.0.0
 repository: https://github.com/Milvintsiss/sqlite3_library_linux
 homepage: https://github.com/Milvintsiss/sqlite3_library_linux
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
